pyproject.toml
[project]
name = "hannibals-army"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "ipykernel>=7.2.0",
    "kagglehub[pandas-datasets]>=1.0.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-tavily>=0.2.17",
    "langchain[anthropic,ollama]>=1.2.10",
    "langgraph>=1.0.10",
    "mcp>=1.26.0",
    "mcp-client>=0.0.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.4",
    "streamlit>=1.54.0",
]
